Add files via upload
Our project aims to create a system that can identify fake news on the internet. We use advanced technology to analyze news articles and websites to determine if the information they provide is true or false. By looking at things like the credibility of the source, the language used, and fact-checking against reliable sources, our system can give a rating on the likelihood that the news is fake. This helps people avoid being misled by false information and promotes more accurate understanding of current events.